orchestrator: Refactor into distinct packages
Before this change, all orchestration-related components were in the
same package, even though there were several independent pieces. They
shared code in sloppy ways, like the global orchestrator calling code in
replicated.go.

This refactors them into multiple pacakages:

- manager/orchestrator/constraintenforcer
- manager/orchestrator/global
- manager/orchestrator/taskreaper
- manager/orchestrator/replicated
- manager/orchestrator/restart
- manager/orchestrator/update

manager/orchestrator has a few shared helper functions, but very little
code stayed there. Also, there is a testutils package for shared testing
functions.

Note that a lot of the tests ended up under
manager/orchestrator/replicated since they depend on the replicated
orchestrator to exercise a different component (such as the task reaper
or restart manager).
