+++ manager/orchestrator/update/updater_test.go
@@ -1,11 +1,12 @@
-package orchestrator
+package update
 
 import (
-	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/docker/swarmkit/api"
+	"github.com/docker/swarmkit/manager/orchestrator"
+	"github.com/docker/swarmkit/manager/orchestrator/restart"
 	"github.com/docker/swarmkit/manager/state"
 	"github.com/docker/swarmkit/manager/state/store"
 	"github.com/docker/swarmkit/protobuf/ptypes"
@@ -14,11 +15,11 @@ import (
 	"golang.org/x/net/context"
 )
 
-func getRunnableSlotSlice(t *testing.T, s *store.MemoryStore, service *api.Service) []slot {
-	runnable, _, err := getRunnableAndDeadSlots(s, service.ID)
+func getRunnableSlotSlice(t *testing.T, s *store.MemoryStore, service *api.Service) []orchestrator.Slot {
+	runnable, _, err := orchestrator.GetRunnableAndDeadSlots(s, service.ID)
 	require.NoError(t, err)
 
-	var runnableSlice []slot
+	var runnableSlice []orchestrator.Slot
 	for _, slot := range runnable {
 		runnableSlice = append(runnableSlice, slot)
 	}
@@ -93,7 +94,7 @@ func TestUpdater(t *testing.T) {
 		assert.NoError(t, store.CreateCluster(tx, cluster))
 		assert.NoError(t, store.CreateService(tx, service))
 		for i := 0; i < instances; i++ {
-			assert.NoError(t, store.CreateTask(tx, newTask(cluster, service, uint64(i), "")))
+			assert.NoError(t, store.CreateTask(tx, orchestrator.NewTask(cluster, service, uint64(i), "")))
 		}
 		return nil
 	})
@@ -109,7 +110,7 @@ func TestUpdater(t *testing.T) {
 
 	service.Spec.Task.GetContainer().Image = "v:2"
 	service.Spec.Task.LogDriver = &api.Driver{Name: "tasklogdriver"}
-	updater := NewUpdater(s, NewRestartSupervisor(s), cluster, service)
+	updater := NewUpdater(s, restart.NewSupervisor(s), cluster, service)
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks := getRunnableSlotSlice(t, s, service)
 	for _, slot := range updatedTasks {
@@ -125,7 +126,7 @@ func TestUpdater(t *testing.T) {
 		Parallelism: 1,
 		Monitor:     ptypes.DurationProto(50 * time.Millisecond),
 	}
-	updater = NewUpdater(s, NewRestartSupervisor(s), cluster, service)
+	updater = NewUpdater(s, restart.NewSupervisor(s), cluster, service)
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks = getRunnableSlotSlice(t, s, service)
 	for _, slot := range updatedTasks {
@@ -142,7 +143,7 @@ func TestUpdater(t *testing.T) {
 		Delay:       *ptypes.DurationProto(10 * time.Millisecond),
 		Monitor:     ptypes.DurationProto(50 * time.Millisecond),
 	}
-	updater = NewUpdater(s, NewRestartSupervisor(s), cluster, service)
+	updater = NewUpdater(s, restart.NewSupervisor(s), cluster, service)
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks = getRunnableSlotSlice(t, s, service)
 	for _, slot := range updatedTasks {
@@ -230,7 +231,7 @@ func TestUpdaterFailureAction(t *testing.T) {
 		assert.NoError(t, store.CreateCluster(tx, cluster))
 		assert.NoError(t, store.CreateService(tx, service))
 		for i := 0; i < instances; i++ {
-			assert.NoError(t, store.CreateTask(tx, newTask(cluster, service, uint64(i), "")))
+			assert.NoError(t, store.CreateTask(tx, orchestrator.NewTask(cluster, service, uint64(i), "")))
 		}
 		return nil
 	})
@@ -244,7 +245,7 @@ func TestUpdaterFailureAction(t *testing.T) {
 	}
 
 	service.Spec.Task.GetContainer().Image = "v:2"
-	updater := NewUpdater(s, NewRestartSupervisor(s), cluster, service)
+	updater := NewUpdater(s, restart.NewSupervisor(s), cluster, service)
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks := getRunnableSlotSlice(t, s, service)
 	v1Counter := 0
@@ -267,7 +268,7 @@ func TestUpdaterFailureAction(t *testing.T) {
 	assert.Equal(t, api.UpdateStatus_PAUSED, service.UpdateStatus.State)
 
 	// Updating again should do nothing while the update is PAUSED
-	updater = NewUpdater(s, NewRestartSupervisor(s), cluster, service)
+	updater = NewUpdater(s, restart.NewSupervisor(s), cluster, service)
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks = getRunnableSlotSlice(t, s, service)
 	v1Counter = 0
@@ -295,7 +296,7 @@ func TestUpdaterFailureAction(t *testing.T) {
 	assert.NoError(t, err)
 
 	service.Spec.Task.GetContainer().Image = "v:3"
-	updater = NewUpdater(s, NewRestartSupervisor(s), cluster, service)
+	updater = NewUpdater(s, restart.NewSupervisor(s), cluster, service)
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks = getRunnableSlotSlice(t, s, service)
 	v2Counter = 0
@@ -374,7 +375,7 @@ func TestUpdaterStopGracePeriod(t *testing.T) {
 	err := s.Update(func(tx store.Tx) error {
 		assert.NoError(t, store.CreateService(tx, service))
 		for i := uint64(0); i < instances; i++ {
-			task := newTask(nil, service, uint64(i), "")
+			task := orchestrator.NewTask(nil, service, uint64(i), "")
 			task.Status.State = api.TaskStateRunning
 			assert.NoError(t, store.CreateTask(tx, task))
 		}
@@ -392,9 +393,9 @@ func TestUpdaterStopGracePeriod(t *testing.T) {
 	before := time.Now()
 
 	service.Spec.Task.GetContainer().Image = "v:2"
-	updater := NewUpdater(s, NewRestartSupervisor(s), nil, service)
+	updater := NewUpdater(s, restart.NewSupervisor(s), nil, service)
 	// Override the default (1 minute) to speed up the test.
-	updater.restarts.taskTimeout = 100 * time.Millisecond
+	updater.restarts.TaskTimeout = 100 * time.Millisecond
 	updater.Run(ctx, getRunnableSlotSlice(t, s, service))
 	updatedTasks := getRunnableSlotSlice(t, s, service)
 	for _, slot := range updatedTasks {
@@ -409,248 +410,5 @@ func TestUpdaterStopGracePeriod(t *testing.T) {
 	// because the system may be slow and it could have taken longer.
 	if after.Sub(before) < 100*time.Millisecond {
 		t.Fatal("stop timeout should have elapsed")
-	}
-}
-
-func TestUpdaterRollback(t *testing.T) {
-	ctx := context.Background()
-	s := store.NewMemoryStore(nil)
-	assert.NotNil(t, s)
-	defer s.Close()
-
-	orchestrator := NewReplicatedOrchestrator(s)
-	defer orchestrator.Stop()
-
-	var (
-		failImage1 uint32
-		failImage2 uint32
-	)
-
-	watchCreate, cancelCreate := state.Watch(s.WatchQueue(), state.EventCreateTask{})
-	defer cancelCreate()
-
-	watchServiceUpdate, cancelServiceUpdate := state.Watch(s.WatchQueue(), state.EventUpdateService{})
-	defer cancelServiceUpdate()
-
-	// Fail new tasks the updater tries to run
-	watchUpdate, cancelUpdate := state.Watch(s.WatchQueue(), state.EventUpdateTask{})
-	defer cancelUpdate()
-	go func() {
-		failedLast := false
-		for {
-			select {
-			case e := <-watchUpdate:
-				task := e.(state.EventUpdateTask).Task
-				if task.DesiredState == task.Status.State {
-					continue
-				}
-				if task.DesiredState == api.TaskStateRunning && task.Status.State != api.TaskStateFailed && task.Status.State != api.TaskStateRunning {
-					err := s.Update(func(tx store.Tx) error {
-						task = store.GetTask(tx, task.ID)
-						// Never fail two image2 tasks in a row, so there's a mix of
-						// failed and successful tasks for the rollback.
-						if task.Spec.GetContainer().Image == "image1" && atomic.LoadUint32(&failImage1) == 1 {
-							task.Status.State = api.TaskStateFailed
-							failedLast = true
-						} else if task.Spec.GetContainer().Image == "image2" && atomic.LoadUint32(&failImage2) == 1 && !failedLast {
-							task.Status.State = api.TaskStateFailed
-							failedLast = true
-						} else {
-							task.Status.State = task.DesiredState
-							failedLast = false
-						}
-						return store.UpdateTask(tx, task)
-					})
-					assert.NoError(t, err)
-				} else if task.DesiredState > api.TaskStateRunning {
-					err := s.Update(func(tx store.Tx) error {
-						task = store.GetTask(tx, task.ID)
-						task.Status.State = task.DesiredState
-						return store.UpdateTask(tx, task)
-					})
-					assert.NoError(t, err)
-				}
-			}
-		}
-	}()
-
-	// Create a service with four replicas specified before the orchestrator
-	// is started. This should result in two tasks when the orchestrator
-	// starts up.
-	err := s.Update(func(tx store.Tx) error {
-		s1 := &api.Service{
-			ID: "id1",
-			Spec: api.ServiceSpec{
-				Annotations: api.Annotations{
-					Name: "name1",
-				},
-				Task: api.TaskSpec{
-					Runtime: &api.TaskSpec_Container{
-						Container: &api.ContainerSpec{
-							Image: "image1",
-						},
-					},
-					Restart: &api.RestartPolicy{
-						Condition: api.RestartOnNone,
-					},
-				},
-				Mode: &api.ServiceSpec_Replicated{
-					Replicated: &api.ReplicatedService{
-						Replicas: 4,
-					},
-				},
-				Update: &api.UpdateConfig{
-					FailureAction:   api.UpdateConfig_ROLLBACK,
-					Parallelism:     1,
-					Delay:           *ptypes.DurationProto(10 * time.Millisecond),
-					Monitor:         ptypes.DurationProto(500 * time.Millisecond),
-					MaxFailureRatio: 0.4,
-				},
-			},
-		}
-
-		assert.NoError(t, store.CreateService(tx, s1))
-		return nil
-	})
-	assert.NoError(t, err)
-
-	// Start the orchestrator.
-	go func() {
-		assert.NoError(t, orchestrator.Run(ctx))
-	}()
-
-	observedTask := watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	atomic.StoreUint32(&failImage2, 1)
-
-	// Start a rolling update
-	err = s.Update(func(tx store.Tx) error {
-		s1 := store.GetService(tx, "id1")
-		require.NotNil(t, s1)
-		s1.PreviousSpec = s1.Spec.Copy()
-		s1.UpdateStatus = nil
-		s1.Spec.Task.GetContainer().Image = "image2"
-		assert.NoError(t, store.UpdateService(tx, s1))
-		return nil
-	})
-	assert.NoError(t, err)
-
-	// Should see three tasks started, then a rollback
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image2")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image2")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image2")
-
-	// Should get to the ROLLBACK_STARTED state
-	for {
-		e := <-watchServiceUpdate
-		if e.(state.EventUpdateService).Service.UpdateStatus == nil {
-			continue
-		}
-		if e.(state.EventUpdateService).Service.UpdateStatus.State == api.UpdateStatus_ROLLBACK_STARTED {
-			break
-		}
-	}
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	// Should end up in ROLLBACK_COMPLETED state
-	for {
-		e := <-watchServiceUpdate
-		if e.(state.EventUpdateService).Service.UpdateStatus.State == api.UpdateStatus_ROLLBACK_COMPLETED {
-			break
-		}
-	}
-
-	atomic.StoreUint32(&failImage1, 1)
-
-	// Repeat the rolling update but this time fail the tasks that the
-	// rollback creates. It should end up in ROLLBACK_PAUSED.
-	err = s.Update(func(tx store.Tx) error {
-		s1 := store.GetService(tx, "id1")
-		require.NotNil(t, s1)
-		s1.PreviousSpec = s1.Spec.Copy()
-		s1.UpdateStatus = nil
-		s1.Spec.Task.GetContainer().Image = "image2"
-		assert.NoError(t, store.UpdateService(tx, s1))
-		return nil
-	})
-	assert.NoError(t, err)
-
-	// Should see three tasks started, then a rollback
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image2")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image2")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image2")
-
-	// Should get to the ROLLBACK_STARTED state
-	for {
-		e := <-watchServiceUpdate
-		if e.(state.EventUpdateService).Service.UpdateStatus == nil {
-			continue
-		}
-		if e.(state.EventUpdateService).Service.UpdateStatus.State == api.UpdateStatus_ROLLBACK_STARTED {
-			break
-		}
-	}
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	observedTask = watchTaskCreate(t, watchCreate)
-	assert.Equal(t, observedTask.Status.State, api.TaskStateNew)
-	assert.Equal(t, observedTask.Spec.GetContainer().Image, "image1")
-
-	// Should end up in ROLLBACK_PAUSED state
-	for {
-		e := <-watchServiceUpdate
-		if e.(state.EventUpdateService).Service.UpdateStatus.State == api.UpdateStatus_ROLLBACK_PAUSED {
-			break
-		}
 	}
 }
